+++ src/components/content/tools/Foam.vue
@@ -180,7 +180,7 @@ import Constants from '@/constants';
 
 import GlobalUtility from '@thzero/library_client/utility/global';
 
-import { useToolsBaseComponent } from '@/components/content/tools/toolBase';
+import { useToolsBaseComponent } from '@/components/content/tools/toolsBase';
 
 import CalculatedOuput from '@/components/content/tools//CalculatedOuput';
 import VFormControl from '@/library_vue_vuetify/components/form/VFormControl';
